[Vulkan] Resource handles are reference counted
This commit makes it possible to have multiple handles (i.e. versions) of a single resource at a time.
This is necessary in some concurrent scenarios, where a previous version of a resource may still be
referenced in an unexecuted command buffer or bound in a descriptor set. In this case, the
introduced reference count indicates so and the disposal is delayed.

Other changes:

- Resource implements IDisposable
- Remove resource-related delete functions (use Dispose() or Disposable.dispose)
- Fix reference counting of device-cached resources

See: https://github.com/aardvark-platform/aardvark.rendering/issues/69