[GL] Move dispatchers to GL.Dispatch
Using GL calls when required extensions are not available leads to
crashes without any output. Our dispatchers either invoke fallback
implementations or throw an exception in this case. This commits moves
those dispatchers from GL to GL.Dispatch. This avoids shadowing with
OpenTK functions, which is dangerous as the order of opening namespaces
determines whether the dispatcher or the OpenTK function is called.
This commits addresses this issue by requiring the dispatcher to be
called explicitly instead.