Improve texture and framebuffer clearing APIs
This commit makes basic clear commands in interfaces use the ClearValues record as argument.
Previously, an overwhelming number of overloads for these interface methods provided flexibility
in how clear values could be defined. With this change, the complexity of definining arbritrary
combinations of clear values is moved to the ClearValues record and its expression builder.
Most of the overloads have been removed, only a handful of frequently used overloads remain, resulting
in a cleaner interface.

Moreover, the API has been extended to allow users to define clear colors for textures with integer formats.