[GL] Fix issues related to framebuffer bindings
This commit fixes some obscure issues with binding and resetting
framebuffers. We usually use the following pattern when working with
temporary framebuffers:

    - Obtain currently bound framebuffer with GL.GetInteger
    - Create, bind and use temporary framebuffer
    - Rebind previously bound framebuffer

Previously, FramebufferTarget.Framebuffer and GetPName.FramebufferBinding
was used with this pattern, which leads to issues when different
framebuffers are bound to the read and draw binding. In that case, the
read binding would be reset to the previous draw framebuffer.