[GL] Disable Dispose for Program
Programs are stored in the cache of the Context
for the lifetime of the context. Deleting a Program
would lead to disposed Program handles remaining
in the cache. As a temporary fix this commit introduces
Program.Free() which properly frees the Program handle
and is called when the cache is disposed. In contrast,
Program.Dispose() is a NOP.

See also: https://github.com/aardvark-platform/aardvark.rendering/commit/dacfd9be6d28c88022dec1a84a24a9e6139f0bb0