[Vulkan] Fix issue with exportable image format queries
vkGetPhysicalDeviceImageFormatProperties2 may fail if a
format is not supported for the given parameters (including
external memory flags). This commit changes GetImageProperties
so that external memory flags are passed only if exporting is
desired.

This fixes a case on an AMD system where the query would fail
even though the format is supported when the not required
external memory flags are omitted.