[GL] Make context sharing more robust
Context sharing may fail if the parent context is in use
when a new context is created. This commit adds a config
flag which causes an additional hidden context to be created.
This context is used as a parent for all other contexts.
That way context sharing will succeed even when all "real"
contexts are busy.