Add discard parameter for buffer upload and copy
The parameter indicates if the implementation may discard the current
content of the destination buffer. This may improve performance in
scenarios where the buffer is used for rendering operations and
updated every frame. On MacOS with OpenGL updating a buffer that is
currently in use will lead to implicit synchronization and massive
stalls. Discarding the current buffer lets OpenGL allocate new
storage and perform the copy operation without waiting for the
rendering operation.

Currently, has no effect in the Vulkan backend. Here, rendering
operations are performed synchronously so that the scenario above
will not occur in the first place.